+++ vue/src/store/modals.js
@@ -1,18 +1,4 @@
-export const randomInteger = (min, max) => {
-  const rand = min + Math.random() * (max - min);
-  return Math.round(rand);
-}
-
-export const genHash = (length = 32) => {
-  let hash = "";
-  const chars = "abcdefghijklmnopqrstuvwxyz0123456789";
-
-  for (let i = 0; i < length; i++) {
-    hash += chars[randomInteger(0, chars.length-1)]
-  }
-
-  return hash + Date.now();
-}
+import { genHash } from "../store/services/common.js"
 
 export default {
   namespaced: true,
